Create component test setup file, and move all shared logic for component tests into it. Move Inbox tests into the tests folder.
